feat: Add metadata tutoriel and simulateur
