+++ src/app/(layout-with-navigation)/(simulation)/simulateur/[root]/_components/Tracking.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import {
   getMatomoEventParcoursTestCategoryStarted,
   matomoEvent50PercentProgress,
