+++ src/app/(layout-with-navigation)/(simulation)/tutoriel/page.tsx
@@ -3,8 +3,15 @@ import ButtonLink from '@/design-system/inputs/ButtonLink'
 import Title from '@/design-system/layout/Title'
 import ButtonStart from './_components/ButtonStart'
 
+import { Metadata } from 'next'
 import AutresQuestions from './_components/AutresQuestions'
 import AvantDeCommencer from './_components/AvantDeCommencer'
+
+export const metadata: Metadata = {
+  title: "Tutoriel du simulateur d'empreinte climat - Nos Gestes Climat",
+  description:
+    'Comprenez comment calculer votre empreinte sur le climat en 10min chrono.',
+}
 
 export default function Tutoriel() {
   return (
@@ -20,12 +27,16 @@ export default function Tutoriel() {
           </>
         }
       />
+
       <AvantDeCommencer />
+
       <AutresQuestions />
+
       <div className="mb-8 flex justify-between border-t border-gray-200 pt-8">
         <ButtonLink href="/" color="secondary">
           ←
         </ButtonLink>
+
         <ButtonStart />
       </div>
     </div>
